+++ sources/useful/ft_getenv.c
@@ -12,63 +12,6 @@
 
 #include "minishell.h"
 
-int	cd_no_file(char **line, char **bis, t_pipe *pipex)
-{
-	if (line[1] && line[1][0] && bis[1] != NULL)
-	{
-		printf("minishell: cd: %s: No such file or directory\n", line[1]);
-		pipex->pipe_ret_b = 1;
-	}
-	ft_split_free(bis);
-	return (1);
-}
-
-int	ft_cd_exec(char **line, char ***envp, char **bis, t_pipe *pipex)
-{
-	int		ret;
-	char	*str;
-
-	pipex->pipe_ret_b = 0;
-	if (ft_get_home(*envp, 1) == NULL)
-		pipex->pipe_ret_b = 1;
-	if (!line[1] || ft_is_str(line[1], "~"))
-		ret = chdir(ft_get_home(*envp, 0));
-	else if (line[1][0] == '~')
-	{
-		str = ft_strjoin(ft_get_home(*envp, 0), line[1] + 1);
-		if (!str)
-			return (2);
-		ret = chdir(str);
-		free(str);
-	}
-	else if (ft_is_str(line[1], "-"))
-		ret = chdir(ft_getenv("OLDPWD", *envp));
-	else
-		ret = chdir(line[1]);
-	if (ret < 0 || bis[1] == NULL)
-		return (cd_no_file(line, bis, pipex));
-	return (0);
-}
-
-int	ft_export_init(char **line, char ***envp, char **tmp)
-{
-	if (!line[1])
-		ft_print_string_of_string(*envp);
-	*tmp = line[1];
-	return (0);
-}
-
-int	ft_unset_i(char **line, char ***envp)
-{
-	int	i;
-
-	i = 0;
-	while (envp[0][i] && \
-	(ft_strncmp(envp[0][i], line[1], ft_strlen_equal(envp[0][i]))) != 0)
-		i++;
-	return (i);
-}
-
 char	*ft_getenv(char *env, char **envp)
 {
 	int	i;
